+++ NEWS.md
@@ -1,5 +1,6 @@
 # qlcData 0.2
 
+* selectively make pies in lmap only for those points that are multi-valued
 * adding options to normalize data in lmap
 * adding option 'ignore.others' in lmap
 * reorder plotting in lmap